requirements.txt
numpy==1.21.0
pandas==1.3.0
torch==1.9.0
transformers==4.9.0
scikit-learn==0.24.2
PyPDF2==1.26.0
pytesseract==0.3.8
opencv-python==4.5.3.20210927
beautifulsoup4==4.10.0
html5lib==1.1
seaborn==0.11.2
matplotlib==3.4.3
tqdm==4.62.3
PyYAML==5.4.1
requests==2.26.0
jupyter==1.0.0
notebook==6.4.3
```